# openapi.yaml
openapi: 3.0.0
paths:
  /api/v1/auth/login:
    post:
      operationId: AuthController_login
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequestDTO"
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponseDTO"
        "401":
          description: Invalid credentials
        "403":
          description: Account banned
      tags: &a1
        - Authentication
    patch:
      operationId: AuthController_changeLogin
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangeLoginRequestDTO"
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChangeLoginResponseDTO"
        "400":
          description: Invalid input
        "401":
          description: Invalid password
        "409":
          description: Login already taken
      security:
        - jwt: []
      tags: *a1
  /api/v1/auth/register:
    post:
      operationId: AuthController_register
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequestDTO"
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegisterResponseDTO"
        "400":
          description: Validation Error
        "403":
          description: Invalid invite token
        "500":
          description: Internal server error
      tags: *a1
  /api/v1/auth/password:
    patch:
      operationId: AuthController_changePassword
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordRequestDTO"
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChangePasswordResponseDTO"
        "400":
          description: Invalid input
        "401":
          description: Invalid current password
      security:
        - jwt: []
      tags: *a1
...
info:
  title: Serchat API
  description: The Serchat API description
  version: "1.0"
  contact: {}
...
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
    LoginRequestDTO:
      type: object
      properties:
        login:
          type: string
        password:
          type: string
      required:
        - login
        - password
    LoginResponseDTO:
      type: object
      properties:
        token:
          type: string
        username:
          type: string
      required:
        - token
        - username
    RegisterRequestDTO:
      type: object
      properties:
        login:
          type: string
        username:
          type: string
        password:
          type: string
        invite:
          type: string
      required:
        - login
        - username
        - password
        - invite
    RegisterResponseDTO:
      type: object
      properties:
        token:
          type: string
      required:
        - token
    ChangeLoginRequestDTO:
      type: object
      properties:
        newLogin:
          type: string
        password:
          type: string
      required:
        - newLogin
        - password
    ChangeLoginResponseDTO:
      type: object
      properties:
        message:
          type: string
        login:
          type: string
        token:
          type: string
      required:
        - message
        - login
        - token
    ChangePasswordRequestDTO:
      type: object
      properties:
        currentPassword:
          type: string
        newPassword:
          type: string
      required:
        - currentPassword
        - newPassword
    ChangePasswordResponseDTO:
      type: object
      properties:
        message:
          type: string
        token:
          type: string
      required:
        - message
        - token
...